[ETFE-3548] - Wipe submission (704) errors when EPC/CN code changed for item
