+++ app/controllers/sections/items/BaseItemsNavigationController.scala
@@ -16,9 +16,10 @@
 
 package controllers.sections.items
 
+import config.Constants.BODYEADESAD
 import controllers.BaseNavigationController
-import models.{GoodsType, Index}
 import models.requests.DataRequest
+import models.{GoodsType, Index, UserAnswers}
 import pages.sections.items._
 import play.api.mvc.Result
 import queries.{ItemsCount, ItemsPackagingCount}
@@ -100,4 +101,39 @@ trait BaseItemsNavigationController extends BaseNavigationController {
         }
       }
     }
+
+  private[controllers] def updateItemSubmissionFailureIndexes(indexOfRemovedItem: Index, userAnswers: UserAnswers): UserAnswers = {
+    val userAnswersWithSubmissionFailureRemovedAtIndex = userAnswers.copy(submissionFailures = userAnswers.submissionFailures.filterNot(
+      _.errorLocation.exists(_.contains(s"$BODYEADESAD[${indexOfRemovedItem.position + 1}]"))
+    ))
+
+    val itemIndexesToAmend: Seq[Int] =
+      userAnswersWithSubmissionFailureRemovedAtIndex.submissionFailures
+        .filter(_.errorLocation.exists(_.contains(BODYEADESAD)))
+        .collect { case itemError =>
+          val lookup = s"$BODYEADESAD\\[(\\d+)\\]".r.unanchored
+          val lookup(index) = itemError.errorLocation.get
+          index.toInt
+        }
+        .filter(_ > (indexOfRemovedItem.position + 1))
+
+    val indexOfSubmissionFailuresNeedingUpdate =
+      itemIndexesToAmend.map { erroredItemIdx =>
+        userAnswersWithSubmissionFailureRemovedAtIndex.submissionFailures.indexWhere(_.errorLocation.exists(_.contains(s"$BODYEADESAD[$erroredItemIdx]"))) -> erroredItemIdx
+      }
+
+    indexOfSubmissionFailuresNeedingUpdate.foldLeft(userAnswersWithSubmissionFailureRemovedAtIndex) {
+      case (userAnswers, (index, erroredIndex)) =>
+
+        val submissionFailureForItem = userAnswers.submissionFailures(index)
+
+        val updatedItem = submissionFailureForItem.copy(errorLocation =
+          submissionFailureForItem.errorLocation.map(
+            _.replace(s"$BODYEADESAD[$erroredIndex]", s"$BODYEADESAD[${erroredIndex - 1}]")
+          )
+        )
+
+        userAnswers.copy(submissionFailures = userAnswers.submissionFailures.updated(index, updatedItem))
+    }
+  }
 }
