+++ test/controllers/sections/items/BaseItemsNavigationControllerSpec.scala
@@ -17,9 +17,9 @@
 package controllers.sections.items
 
 import base.SpecBase
-import fixtures.ItemFixtures
+import fixtures.{ItemFixtures, MovementSubmissionFailureFixtures}
 import mocks.services.{MockGetCnCodeInformationService, MockUserAnswersService}
-import models.UserAnswers
+import models.{Index, UserAnswers}
 import models.requests.DataRequest
 import models.response.referenceData.{BulkPackagingType, CnCodeInformation, ItemPackaging}
 import models.sections.items.ItemBulkPackagingCode.BulkSolidPowders
@@ -36,7 +36,8 @@ import scala.concurrent.Future
 class BaseItemsNavigationControllerSpec extends SpecBase
   with MockGetCnCodeInformationService
   with MockUserAnswersService
-  with ItemFixtures {
+  with ItemFixtures
+  with MovementSubmissionFailureFixtures {
 
   class Test(val userAnswers: UserAnswers) {
     implicit val request: DataRequest[_] = dataRequest(FakeRequest(), userAnswers)
@@ -109,6 +110,119 @@ class BaseItemsNavigationControllerSpec extends SpecBase
 
         status(result) mustBe SEE_OTHER
         redirectLocation(result).value mustBe routes.ItemsIndexController.onPageLoad(request.ern, request.draftId).url
+      }
+    }
+  }
+
+  //scalastyle:off
+  ".updateItemSubmissionFailureIndexes" - {
+
+    "when multiple errors for multiple items exists" - {
+
+      val userAnswersWithFailures = emptyUserAnswers.copy(submissionFailures = Seq(
+        movementSubmissionFailure,
+        itemQuantityFailure(1),
+        itemQuantityFailure(4),
+        itemQuantityFailure(10),
+        movementSubmissionFailure,
+        itemQuantityFailure(20),
+        itemQuantityFailure(11)
+      ))
+
+      "when removing an item at the start (first item with error)" - {
+
+        "should remove the item failure AND reindex XPath for ErrorLocations to synchronise them with the new array order" in new Test(userAnswersWithFailures) {
+
+          val result = controller.updateItemSubmissionFailureIndexes(Index(0), userAnswersWithFailures)
+
+          result.submissionFailures mustBe Seq(
+            movementSubmissionFailure,
+            itemQuantityFailure(3),
+            itemQuantityFailure(9),
+            movementSubmissionFailure,
+            itemQuantityFailure(19),
+            itemQuantityFailure(10)
+          )
+        }
+      }
+
+      "when removing an item at a mid-point" - {
+
+        "should remove the item failure AND reindex XPath for ErrorLocations to synchronise them with the new array order" in new Test(userAnswersWithFailures) {
+
+          val result = controller.updateItemSubmissionFailureIndexes(Index(3), userAnswersWithFailures)
+
+          result.submissionFailures mustBe Seq(
+            movementSubmissionFailure,
+            itemQuantityFailure(1),
+            itemQuantityFailure(9),
+            movementSubmissionFailure,
+            itemQuantityFailure(19),
+            itemQuantityFailure(10)
+          )
+        }
+      }
+
+      "when removing an item at the end (last item with error)" - {
+
+        "should remove the item failure AND reindex XPath for ErrorLocations to synchronise them with the new array order" in new Test(userAnswersWithFailures) {
+
+          val result = controller.updateItemSubmissionFailureIndexes(Index(19), userAnswersWithFailures)
+
+          result.submissionFailures mustBe Seq(
+            movementSubmissionFailure,
+            itemQuantityFailure(1),
+            itemQuantityFailure(4),
+            itemQuantityFailure(10),
+            movementSubmissionFailure,
+            itemQuantityFailure(11)
+          )
+        }
+      }
+
+      "when removing an item that doesn't have an error against it" - {
+
+        "should update all subsequent XPath for ErrorLocations to synchronise them with the new array order" in new Test(userAnswersWithFailures) {
+
+          val result = controller.updateItemSubmissionFailureIndexes(Index(15), userAnswersWithFailures)
+
+          result.submissionFailures mustBe Seq(
+            movementSubmissionFailure,
+            itemQuantityFailure(1),
+            itemQuantityFailure(4),
+            itemQuantityFailure(10),
+            movementSubmissionFailure,
+            itemQuantityFailure(19),
+            itemQuantityFailure(11)
+          )
+        }
+      }
+
+      "when removing an item and not submission failures exist for any items" - {
+
+        val userAnswersWithFailures = emptyUserAnswers.copy(submissionFailures = Seq(movementSubmissionFailure))
+
+        "should return the submission failures unaffected" in new Test(userAnswersWithFailures) {
+
+          val result = controller.updateItemSubmissionFailureIndexes(Index(15), userAnswersWithFailures)
+
+          result.submissionFailures mustBe Seq(movementSubmissionFailure)
+        }
+      }
+
+      "when removing an item before another item" - {
+
+        val userAnswersWithFailures = emptyUserAnswers.copy(submissionFailures = Seq(
+          itemQuantityFailure(1).copy(originalAttributeValue = Some("1")),
+          itemQuantityFailure(2)
+        ))
+
+        "should remove the item failure AND reindex XPath for ErrorLocations to synchronise them with the new array order" in new Test(userAnswersWithFailures) {
+
+          val result = controller.updateItemSubmissionFailureIndexes(Index(0), userAnswersWithFailures)
+
+          result.submissionFailures mustBe Seq(itemQuantityFailure(1))
+        }
       }
     }
   }
