+++ test/controllers/sections/items/ItemRemoveItemControllerSpec.scala
@@ -21,7 +21,7 @@ import controllers.actions.FakeDataRetrievalAction
 import fixtures.{ItemFixtures, MovementSubmissionFailureFixtures}
 import forms.sections.items.ItemRemoveItemFormProvider
 import mocks.services.MockUserAnswersService
-import models.{Index, UserAnswers}
+import models.UserAnswers
 import navigation.FakeNavigators.FakeItemsNavigator
 import pages.sections.items.ItemExciseProductCodePage
 import play.api.mvc.{AnyContentAsEmpty, Call}
@@ -137,118 +137,6 @@ class ItemRemoveItemControllerSpec extends SpecBase with MockUserAnswersService 
 
       status(result) mustEqual SEE_OTHER
       redirectLocation(result).value mustEqual controllers.routes.JourneyRecoveryController.onPageLoad().url
-    }
-  }
-
-  ".updateItemSubmissionFailureIndexes" - {
-
-    "when multiple errors for multiple items exists" - {
-
-      val userAnswersWithFailures = emptyUserAnswers.copy(submissionFailures = Seq(
-        movementSubmissionFailure,
-        itemQuantityFailure(1),
-        itemQuantityFailure(4),
-        itemQuantityFailure(10),
-        movementSubmissionFailure,
-        itemQuantityFailure(20),
-        itemQuantityFailure(11)
-      ))
-
-      "when removing an item at the start (first item with error)" - {
-
-        "should remove the item failure AND reindex XPath for ErrorLocations to synchronise them with the new array order" in new Test(Some(userAnswersWithFailures)) {
-
-          val result = controller.updateItemSubmissionFailureIndexes(Index(0), userAnswersWithFailures)
-
-          result.submissionFailures mustBe Seq(
-            movementSubmissionFailure,
-            itemQuantityFailure(3),
-            itemQuantityFailure(9),
-            movementSubmissionFailure,
-            itemQuantityFailure(19),
-            itemQuantityFailure(10)
-          )
-        }
-      }
-
-      "when removing an item at a mid-point" - {
-
-        "should remove the item failure AND reindex XPath for ErrorLocations to synchronise them with the new array order" in new Test(Some(userAnswersWithFailures)) {
-
-          val result = controller.updateItemSubmissionFailureIndexes(Index(3), userAnswersWithFailures)
-
-          result.submissionFailures mustBe Seq(
-            movementSubmissionFailure,
-            itemQuantityFailure(1),
-            itemQuantityFailure(9),
-            movementSubmissionFailure,
-            itemQuantityFailure(19),
-            itemQuantityFailure(10)
-          )
-        }
-      }
-
-      "when removing an item at the end (last item with error)" - {
-
-        "should remove the item failure AND reindex XPath for ErrorLocations to synchronise them with the new array order" in new Test(Some(userAnswersWithFailures)) {
-
-          val result = controller.updateItemSubmissionFailureIndexes(Index(19), userAnswersWithFailures)
-
-          result.submissionFailures mustBe Seq(
-            movementSubmissionFailure,
-            itemQuantityFailure(1),
-            itemQuantityFailure(4),
-            itemQuantityFailure(10),
-            movementSubmissionFailure,
-            itemQuantityFailure(11)
-          )
-        }
-      }
-
-      "when removing an item that doesn't have an error against it" - {
-
-        "should update all subsequent XPath for ErrorLocations to synchronise them with the new array order" in new Test(Some(userAnswersWithFailures)) {
-
-          val result = controller.updateItemSubmissionFailureIndexes(Index(15), userAnswersWithFailures)
-
-          result.submissionFailures mustBe Seq(
-            movementSubmissionFailure,
-            itemQuantityFailure(1),
-            itemQuantityFailure(4),
-            itemQuantityFailure(10),
-            movementSubmissionFailure,
-            itemQuantityFailure(19),
-            itemQuantityFailure(11)
-          )
-        }
-      }
-
-      "when removing an item and not submission failures exist for any items" - {
-
-        val userAnswersWithFailures = emptyUserAnswers.copy(submissionFailures = Seq(movementSubmissionFailure))
-
-        "should return the submission failures unaffected" in new Test(Some(userAnswersWithFailures)) {
-
-          val result = controller.updateItemSubmissionFailureIndexes(Index(15), userAnswersWithFailures)
-
-          result.submissionFailures mustBe Seq(movementSubmissionFailure)
-        }
-      }
-
-      "when removing an item before another item" - {
-
-        val userAnswersWithFailures = emptyUserAnswers.copy(submissionFailures = Seq(
-          itemQuantityFailure(1).copy(originalAttributeValue = Some("1")),
-          itemQuantityFailure(2)
-        ))
-
-        "should remove the item failure AND reindex XPath for ErrorLocations to synchronise them with the new array order" in new Test(Some(userAnswersWithFailures)) {
-
-          val result = controller.updateItemSubmissionFailureIndexes(Index(0), userAnswersWithFailures)
-
-          result.submissionFailures mustBe Seq(itemQuantityFailure(1))
-        }
-      }
     }
   }
 }
