+++ test/mocks/services/MockUserAnswersService.scala
@@ -40,7 +40,8 @@ trait MockUserAnswersService extends MockFactory {
         .expects(where { (actualAnswers, _) =>
           actualAnswers.ern == userAnswers.ern &&
             //Declaration page stores a LocalDateTime which can't be asserted against reliably
-            actualAnswers.data - DeclarationPage == userAnswers.data - DeclarationPage
+            actualAnswers.data - DeclarationPage == userAnswers.data - DeclarationPage &&
+            actualAnswers.submissionFailures == userAnswers.submissionFailures
         })
 
     def set(): CallHandler2[UserAnswers, HeaderCarrier, Future[UserAnswers]] =
